use text size in code
