+++ lib/theme/fonts.gen.dart
@@ -17,21 +17,3 @@ class FontFamily {
   static const String roboto = 'Roboto';
 }
 
-class TextSize {
-  TextSize._();
-
-  /// Text size for Page title
-  static const double pageTitle = 24.0;
-
-  /// Text size for News title
-  static const double newsTitle = 20.0;
-
-  /// Text size for Medium title
-  static const double mediumTitle = 16.0;
-
-  /// Text size for Small title
-  static const double smallTitle = 12.0;
-
-  /// Text size for Description
-  static const double description = 11.0;
-}
