+++ lib/ui/menu_item_list_view.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:pola_flutter/theme/fonts.gen.dart';
+import 'package:pola_flutter/theme/text_size.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:pola_flutter/analytics/analytics_about_row.dart';
 import 'package:pola_flutter/analytics/pola_analytics.dart';
@@ -135,7 +136,7 @@ class _MenuBottomItem extends StatelessWidget {
   });
 
   final textStyle = const TextStyle(
-      fontWeight: FontWeight.w500, fontSize: 16.0, fontFamily: FontFamily.lato);
+      fontWeight: FontWeight.w500, fontSize: TextSize.mediumTitle, fontFamily: FontFamily.lato);
 
   @override
   Widget build(BuildContext context) {
